Split of branch guessing from environment
